refactor: implement granular suspense for dashboard
- Convert dashboard sections to Server Components
- Implement granular Suspense boundaries with Skeletons
- Extract client logic to separate components
- Remove page-level Suspense and monolithic Skeleton
- Fix CSS media query syntax and align skeleton styles
