+++ src/app/(main)/dashboard/page.tsx
@@ -1,5 +1,5 @@
 import { Metadata } from "next";
-import { Suspense } from "react";
+
 import { getPageMetadata } from "@/config/metadata";
 import styles from "./DashboardPage.module.css";
 import * as Dashboard from "@/features/dashboard/components";
@@ -11,9 +11,7 @@ const DashboardPage = () => {
 		<>
 			<h1 className={`${styles["dashboard-title"]}`}>ダッシュボード</h1>
 			<div className={styles["dashboard-content-wrapper"]}>
-				<Suspense fallback={<Dashboard.DashboardActivitySkeleton />}>
-					<Dashboard.DashboardContent />
-				</Suspense>
+				<Dashboard.DashboardContent />
 			</div>
 		</>
 	);
