+++ src/features/dashboard/components/dashboard-activity-skeleton/DashboardActivitySkeleton.module.css
@@ -16,7 +16,7 @@
 	color: var(--color-primary-white);
 	text-shadow: 1.5px 1.5px 0px var(--color-primary-black);
 
-	@media (width < 1024px) {
+	@media (max-width: 1023px) {
 		place-self: center;
 		font-size: 1.125rem;
 	}
@@ -34,7 +34,7 @@
 	padding-block: 0 10px;
 	margin-block-start: 20px;
 
-	@media (width < 768px) {
+	@media (max-width: 767px) {
 		padding-inline: 0;
 	}
 }
@@ -45,10 +45,10 @@
 	display: block flex;
 	justify-content: space-between;
 	align-items: center;
-	border: 7px dashed var(--color-primary-brown-light);
-	padding-inline: 5px;
-	padding-block: 5px;
-	border-radius: var(--border-radius-small);
+	border: 3mm ridge #b69973;
+	padding-inline: 2.5px;
+	padding-block: 2.5px;
+	/* border-radius: var(--border-radius-small); */
 	background-color: var(--color-primary-brown-dark);
 }
 
@@ -57,7 +57,7 @@
 	width: 100%;
 	display: block grid;
 	color: var(--color-primary-white);
-	background-color: var(--color-primary-brown);
+	background-color: #ae926e;
 	border: 2px solid var(--color-primary-brown-light);
 	padding-inline: 15px;
 	padding-block: 12px 10px;
@@ -79,7 +79,7 @@
 	animation: shimmer 1.5s infinite;
 	border-radius: 4px;
 
-	@media (width < 1024px) {
+	@media (max-width: 1023px) {
 		height: 24px;
 	}
 }
