+++ src/features/dashboard/components/dashboard-content/DashboardContent.module.css
@@ -1,5 +1,5 @@
 .dashboard-content-container {
-	display: block grid;
+	display: grid;
 	grid-template-columns: auto;
 	grid-auto-rows: minmax(min-content, max-content);
 	gap: 25px;
@@ -11,13 +11,13 @@
 
 .dashboard-zenn-area {
 	width: 100%;
-	display: block grid;
+	display: grid;
 	grid-template-columns: auto 1fr;
 	gap: 35px;
 	padding-inline: 20px;
 	padding-block: 20px;
 
-	@media (width < 768px) {
+	@media (max-width: 767px) {
 		grid-template-columns: auto;
 	}
 }
