+++ src/features/dashboard/components/dashboard-content/DashboardContent.tsx
@@ -1,41 +1,61 @@
 import { Suspense } from "react";
-import { getDashboardHeroData } from "@/features/dashboard/_lib/fetcher";
-import { heroLevelAndItemRelation } from "@/features/items/data/itemsData";
-import * as Dashboard from "@/features/dashboard/components";
+import styles from "./DashboardContent.module.css";
+
+// Components
+import DashboardHeroSection from "../dashboard-hero-section/DashboardHeroSection";
+import DashboardPlatformStatsSection from "../dashboard-platform-stats-section/DashboardPlatformStatsSection";
+import DashboardActivitySection from "../dashboard-activity-section/DashboardActivitySection";
+import DashboardLatestPartyMemberSection from "../dashboard-latest-party-member-section/DashboardLatestPartyMemberSection";
+import DashboardLatestItemSection from "../dashboard-latest-item-section/DashboardLatestItemSection";
+
+// Skeletons
+import DashboardHeroSkeleton from "../dashboard-hero-skeleton/DashboardHeroSkeleton";
+import DashboardPlatformStatsSkeleton from "../dashboard-platform-stats-skeleton/DashboardPlatformStatsSkeleton";
+import DashboardActivitySkeleton from "../dashboard-activity-skeleton/DashboardActivitySkeleton";
+import DashboardLatestPartyMemberSkeleton from "../dashboard-latest-party-member-skeleton/DashboardLatestPartyMemberSkeleton";
+import DashboardLatestItemSkeleton from "../dashboard-latest-item-skeleton/DashboardLatestItemSkeleton";
 
 /**
  * DashboardContent (Server Component)
  *
- * Request Memoizationを活用したデータフェッチを行い、
- * 取得したデータをClient Componentに渡す。
- *
- * Server ComponentとClient Componentを適切に配置する。
- *
- * 変更点:
- * - Client Component → Server Componentに変換
- * - useHero(), useState, useEffectを削除
- * - データフェッチをサーバー側で同期的に実行
- * - Server ComponentはここでレンダリングしてClient Componentにslotとして渡す
+ * Granular Suspense Pattern:
+ * 各セクションを独立したSuspense境界で囲み、
+ * 個別にデータフェッチとローディング状態を管理する。
+ * これにより、ユーザーは準備ができた部分から順次コンテンツを見ることができる。
  */
-const DashboardContent = async () => {
-	// Request Memoization: 同一リクエスト内で1回のみ実行
-	const heroData = await getDashboardHeroData();
-
-	// lastAcquiredItemIdを計算（useEffectのロジックを移植）
-	const acquiredItemIds = Object.entries(heroLevelAndItemRelation)
-		.filter(([, requiredLevel]) => heroData.level >= requiredLevel)
-		.map(([itemIdStr]) => parseInt(itemIdStr, 10));
-
-	const lastAcquiredItemId = acquiredItemIds.length > 0 ? Math.max(...acquiredItemIds) : null;
-
-	// Client Componentにデータを渡す
-	// Server ComponentはCompositionパターン（children）で渡す
+const DashboardContent = () => {
 	return (
-		<Dashboard.DashboardContentClient heroData={heroData} lastAcquiredItemId={lastAcquiredItemId}>
-			<Suspense fallback={<Dashboard.DashboardActivitySkeleton />}>
-				<Dashboard.DashboardActivitySection />
+		<div className={styles["dashboard-content-container"]}>
+			<Suspense fallback={<DashboardHeroSkeleton />}>
+				<DashboardHeroSection />
 			</Suspense>
-		</Dashboard.DashboardContentClient>
+
+			<hr />
+
+			<div className={styles["dashboard-zenn-area"]}>
+				<Suspense fallback={<DashboardPlatformStatsSkeleton />}>
+					<DashboardPlatformStatsSection />
+				</Suspense>
+
+				<hr className="block md:hidden" />
+
+				<Suspense fallback={<DashboardActivitySkeleton />}>
+					<DashboardActivitySection />
+				</Suspense>
+			</div>
+
+			<hr />
+
+			<Suspense fallback={<DashboardLatestPartyMemberSkeleton />}>
+				<DashboardLatestPartyMemberSection />
+			</Suspense>
+
+			<hr />
+
+			<Suspense fallback={<DashboardLatestItemSkeleton />}>
+				<DashboardLatestItemSection />
+			</Suspense>
+		</div>
 	);
 };
