+++ src/features/dashboard/components/dashboard-latest-party-member-skeleton/DashboardLatestPartyMemberSkeleton.module.css
@@ -1,82 +1,69 @@
-/* Skeleton Box - matches party-member-box */
-.skeleton-box {
-	display: block grid;
-	background-color: var(--color-primary-brown-dark);
-	border: 7px dashed var(--color-primary-brown-light);
-	padding-inline: 5px;
-	padding-block: 5px;
-	border-radius: var(--border-radius-small);
+/* matches latest-party-member-section (and latest-item-section) */
+.skeleton-latest-section {
+	width: 100%;
+	display: grid;
+	grid-template-rows: auto 1fr;
+	gap: 15px;
 }
 
-/* Skeleton Link - matches party-member-link */
-.skeleton-link {
-	display: block grid;
-	grid-template-columns: auto 1fr;
-	gap: 25px;
-	background-color: var(--color-primary-brown);
-	border: 2px solid var(--color-primary-brown-light);
-	padding-block: 15px;
-	padding-inline: 20px;
-
-	@media (width < 640px) {
-		grid-template-columns: 1fr;
-		gap: 10px;
-	}
-}
-
-/* Skeleton Icon */
-.skeleton-icon {
-	width: 61px;
-	height: 61px;
+/* matches latest-party-member-title */
+.skeleton-section-title {
+	width: 200px;
+	height: 28px;
 	background: linear-gradient(90deg, #2a2a2a 25%, #3a3a3a 50%, #2a2a2a 75%);
 	background-size: 200% 100%;
 	animation: shimmer 1.5s infinite;
 	border-radius: 4px;
-	margin-inline: auto;
+	margin-block-end: 15px;
+
+	@media (max-width: 1023px) {
+		place-self: center;
+	}
 }
 
-/* Skeleton Info - matches party-member-info */
-.skeleton-info {
-	display: block grid;
+/* matches latest-party-member-container */
+.skeleton-party-member-container {
+	width: 100%;
+	display: grid;
+	place-self: start;
 	gap: 10px;
-	padding-block: 5px;
+	margin-block-start: 20px;
+	padding-inline: 25px;
 
-	@media (width < 640px) {
+	@media (max-width: 767px) {
 		place-items: center;
+		gap: 15px;
+		padding-inline: 0;
 	}
 }
 
-/* Skeleton Name */
-.skeleton-name {
-	width: 120px;
-	height: 100%;
+/* matches latest-party-member-card/guest-message */
+.skeleton-party-member-guest {
+	min-width: 300px;
+	height: 140px;
+	background: linear-gradient(90deg, #2a2a2a 25%, #3a3a3a 50%, #2a2a2a 75%);
+	background-size: 200% 100%;
+	animation: shimmer 1.5s infinite;
+	/* border: 7px dashed var(--color-primary-brown-light); */
+	border: 2.7mm ridge #b69973;
+	background-color: var(--color-primary-brown-dark);
+	padding-inline: 2.5px;
+	padding-block: 2.5px;
+	/* border-radius: var(--border-radius-small); */
+}
+
+/* matches latest-party-member-share-link-text */
+.skeleton-party-share {
+	width: 200px;
+	height: 20px;
 	background: linear-gradient(90deg, #2a2a2a 25%, #3a3a3a 50%, #2a2a2a 75%);
 	background-size: 200% 100%;
 	animation: shimmer 1.5s infinite;
 	border-radius: 4px;
-
-	@media (width < 640px) {
-		width: 60%;
-		height: 20px;
-	}
+	margin-top: 10px;
+	place-self: center; /* Center share link skeleton */
 }
 
-/* Skeleton Description */
-.skeleton-description {
-	width: 250px;
-	height: 100%;
-	background: linear-gradient(90deg, #2a2a2a 25%, #3a3a3a 50%, #2a2a2a 75%);
-	background-size: 200% 100%;
-	animation: shimmer 1.5s infinite;
-	border-radius: 4px;
-
-	@media (width < 640px) {
-		width: 100%;
-		height: 20px;
-	}
-}
-
-/* Shimmer Animation */
 @keyframes shimmer {
 	0% {
 		background-position: 200% 0;
