+++ src/features/dashboard/components/dashboard-platform-stats-section/DashboardPlatformStatsSection.tsx
@@ -1,61 +1,20 @@
-"use client";
-
-import { useState, useEffect } from "react";
 import styles from "./DashboardPlatformStatsSection.module.css";
-import { DashboardData } from "../../types/dashboard.types";
-import { useHero } from "@/contexts/HeroContext";
-import { getUserInfo } from "@/lib/api/user";
-import { useUser } from "@clerk/nextjs";
 import XShareButton from "@/components/common/x-share-button/XShareButton";
 import Image from "next/image";
+import { getDashboardHeroData } from "@/features/dashboard/_lib/fetcher";
+import { getUser } from "@/features/user/_lib/fetcher";
 
-type DashboardPlatformStatsSectionProps = {
-	dashboardData: DashboardData;
-};
+const DashboardPlatformStatsSection = async () => {
+	// Request Memoizationにより、他コンポーネントと同じフェッチを共有
+	const [heroData, user] = await Promise.all([getDashboardHeroData(), getUser()]);
 
-const DashboardPlatformStatsSection = ({ dashboardData }: DashboardPlatformStatsSectionProps) => {
-	const { heroData, isLoading, error } = useHero();
-	const { user, isLoaded } = useUser();
-	const [userZennInfo, setUserZennInfo] = useState<{
-		zennUsername?: string;
-	} | null>(null);
-
-	// ゲストユーザーの判定
-	const isGuestUser = !isLoaded || !user || !userZennInfo?.zennUsername;
-
-	// ユーザーのZenn連携情報を取得
-	useEffect(() => {
-		const fetchUserZennInfo = async () => {
-			if (!isLoaded) {
-				return;
-			}
-
-			if (!user) {
-				setUserZennInfo(null);
-				return;
-			}
-
-			try {
-				const userData = await getUserInfo();
-
-				if (userData.success && userData.user) {
-					setUserZennInfo(userData.user);
-				} else {
-					setUserZennInfo(null);
-				}
-			} catch (err) {
-				console.error("ユーザー情報取得エラー:", err);
-				setUserZennInfo(null);
-			}
-		};
-
-		fetchUserZennInfo();
-	}, [isLoaded, user?.id]);
+	// ゲストユーザーかどうかの判定
+	const isGuestUser = !user || !user.zennUsername;
 
 	// 実際のZennデータとモックデータを組み合わせてstatsを作成
 	const zennStat = {
 		platform: "Zenn",
-		count: isLoading ? dashboardData.postStats[0].count : heroData.level,
+		count: heroData.level,
 		color: "#3ea8ff",
 	};
 
@@ -77,26 +36,8 @@ const DashboardPlatformStatsSection = ({ dashboardData }: DashboardPlatformStats
 						<div className={`${styles["platform-stat-card-content"]}`}>
 							<h3 className={`${styles["platform-stat-card-title"]}`}>{zennStat.platform}</h3>
 							<div className={`${styles["platform-stat-count"]}`}>
-								{isLoading ? (
-									<>
-										<span className={styles["loading-dots"]}>
-											<span className={styles["loading-dot"]}>.</span>
-											<span className={styles["loading-dot"]}>.</span>
-											<span className={styles["loading-dot"]}>.</span>
-										</span>
-										<span className={`${styles["platform-stat-unit"]}`}>記事</span>
-									</>
-								) : error ? (
-									<>
-										<span className={`${styles["platform-stat-error"]}`}>0</span>
-										<span className={`${styles["platform-stat-unit"]}`}>記事</span>
-									</>
-								) : (
-									<>
-										<em className={`${styles["platform-stat-count-em"]}`}>{zennStat.count}</em>
-										<span className={`${styles["platform-stat-unit"]}`}>記事</span>
-									</>
-								)}
+								<em className={`${styles["platform-stat-count-em"]}`}>{zennStat.count}</em>
+								<span className={`${styles["platform-stat-unit"]}`}>記事</span>
 							</div>
 						</div>
 					</div>
