+++ src/features/dashboard/components/index.ts
@@ -2,26 +2,26 @@ import DashboardActivityContent from "./dashboard-activity-content/DashboardActi
 import DashboardActivitySection from "./dashboard-activity-section/DashboardActivitySection";
 import DashboardActivitySkeleton from "./dashboard-activity-skeleton/DashboardActivitySkeleton";
 import DashboardContent from "./dashboard-content/DashboardContent";
-import DashboardContentClient from "./dashboard-content-client/DashboardContentClient";
 import DashboardHeroSection from "./dashboard-hero-section/DashboardHeroSection";
+import DashboardHeroSkeleton from "./dashboard-hero-skeleton/DashboardHeroSkeleton";
 import DashboardLatestItemSection from "./dashboard-latest-item-section/DashboardLatestItemSection";
 import DashboardLatestItemSkeleton from "./dashboard-latest-item-skeleton/DashboardLatestItemSkeleton";
 import DashboardLatestPartyMemberSection from "./dashboard-latest-party-member-section/DashboardLatestPartyMemberSection";
 import DashboardLatestPartyMemberSkeleton from "./dashboard-latest-party-member-skeleton/DashboardLatestPartyMemberSkeleton";
 import DashboardPlatformStatsSection from "./dashboard-platform-stats-section/DashboardPlatformStatsSection";
-import DashboardSkeleton from "./dashboard-skeleton/DashboardSkeleton";
+import DashboardPlatformStatsSkeleton from "./dashboard-platform-stats-skeleton/DashboardPlatformStatsSkeleton";
 
 export {
 	DashboardActivityContent,
 	DashboardActivitySection,
 	DashboardActivitySkeleton,
 	DashboardContent,
-	DashboardContentClient,
 	DashboardHeroSection,
+	DashboardHeroSkeleton,
 	DashboardLatestItemSection,
 	DashboardLatestItemSkeleton,
 	DashboardLatestPartyMemberSection,
 	DashboardLatestPartyMemberSkeleton,
 	DashboardPlatformStatsSection,
-	DashboardSkeleton,
+	DashboardPlatformStatsSkeleton,
 };
